third_party/blink/web_tests/animations/keyframe-selector-negative-percentage.html

<!DOCTYPE html>
<html>
<head>
<title>Check that negative (out of range) key frame percentage doesn't cause subsequent first or second ruleset to be skipped.</title>
<style>
#test { background-color: red; color: black }
@keyframes fade { -10% { color: red; } }
#test { background-color: green; }                 /* if skipped, then background will be red */
#test { color: white; }                            /* if skipped, then text will be black */
</style>
</head>
<body>
<div id="test">The background color of this element should be green and the text should be white.</div>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
var r = '';
var s = getComputedStyle(document.getElementById("test"));
assert_equals(s.backgroundColor, "rgb(0, 128, 0)", 'background color should be green');
assert_equals(s.color, "rgb(255, 255, 255)", 'text color should be white');
done();
</script>
</body>
</html>
